bịp

+++ DoAnTTTT_QuanLyCuaHieuCamDo/DoAnTTTT_QuanLyCuaHieuCamDo/DTO/ChiTietHoaDonCamDTO.cs
@@ -12,8 +12,8 @@ namespace DoAnTTTT_QuanLyCuaHieuCamDo.DTO
         private string maSP;
         private string tenLoai;
         private string tenSP;
-        private float dinhGia;
-        private float giaThanhLy;
+        private double dinhGia;
+        private double giaThanhLy;
         private string moTa;
         private string mauSac;
         private string hienTrang;
@@ -23,15 +23,15 @@ namespace DoAnTTTT_QuanLyCuaHieuCamDo.DTO
         public string MaSP { get => maSP; set => maSP = value; }
         public string TenLoai { get => tenLoai; set => tenLoai = value; }
         public string TenSP { get => tenSP; set => tenSP = value; }
-        public float DinhGia { get => dinhGia; set => dinhGia = value; }
-        public float GiaThanhLy { get => giaThanhLy; set => giaThanhLy = value; }
+        public double DinhGia { get => dinhGia; set => dinhGia = value; }
+        public double GiaThanhLy { get => giaThanhLy; set => giaThanhLy = value; }
         public string MoTa { get => moTa; set => moTa = value; }
         public string MauSac { get => mauSac; set => mauSac = value; }
         public string HienTrang { get => hienTrang; set => hienTrang = value; }
         public string NhanHieu { get => nhanHieu; set => nhanHieu = value; }
 
 
-        public ChiTietHoaDonCamDTO(int MaHoaDonCam, string MaSP, string TenLoai, string TenSP, float DinhGia, float GiaThanhLy, string MoTa, string MauSac, string HienTrang, string NhangHieu)
+        public ChiTietHoaDonCamDTO(int MaHoaDonCam, string MaSP, string TenLoai, string TenSP, double DinhGia, double GiaThanhLy, string MoTa, string MauSac, string HienTrang, string NhangHieu)
         {
             this.MaHoaDonCam = MaHoaDonCam;
             this.MaSP = MaSP;
@@ -51,8 +51,8 @@ namespace DoAnTTTT_QuanLyCuaHieuCamDo.DTO
             this.MaSP = row["MaSP"].ToString();
             this.TenLoai = row["TenLoai"].ToString();
             this.TenSP = row["TenSP"].ToString();
-            this.DinhGia = (float)Convert.ToDouble(row["DinhGia"]);
-            this.GiaThanhLy = (float)Convert.ToDouble(row["GiaThanhLy"]);
+            this.DinhGia = (double)Convert.ToDouble(row["DinhGia"]);
+            this.GiaThanhLy = (double)Convert.ToDouble(row["GiaThanhLy"]);
             this.MoTa = row["MoTa"].ToString();
             this.MauSac = row["MauSac"].ToString();
             this.HienTrang = row["HienTrang"].ToString();
