+++ DoAnTTTT_QuanLyCuaHieuCamDo/DoAnTTTT_QuanLyCuaHieuCamDo/DTO/SanPhamDTO.cs
@@ -11,8 +11,8 @@ namespace DoAnTTTT_QuanLyCuaHieuCamDo.DTO
         private string maSP;
         private string tenLoai;
         private string tenSP;
-        private float dinhGia;
-        private float giaThanhLy;
+        private double dinhGia;
+        private double giaThanhLy;
         private string moTa;
         private string mauSac;
         private string hienTrang;
@@ -25,8 +25,8 @@ namespace DoAnTTTT_QuanLyCuaHieuCamDo.DTO
         public string MaSP { get => maSP; set => maSP = value; }
         public string TenLoai { get => tenLoai; set => tenLoai = value; }
         public string TenSP { get => tenSP; set => tenSP = value; }
-        public float DinhGia { get => dinhGia; set => dinhGia = value; }
-        public float GiaThanhLy { get => giaThanhLy; set => giaThanhLy = value; }
+        public double DinhGia { get => dinhGia; set => dinhGia = value; }
+        public double GiaThanhLy { get => giaThanhLy; set => giaThanhLy = value; }
         public string MoTa { get => moTa; set => moTa = value; }
         public string MauSac { get => mauSac; set => mauSac = value; }
         public string HienTrang { get => hienTrang; set => hienTrang = value; }
@@ -37,7 +37,7 @@ namespace DoAnTTTT_QuanLyCuaHieuCamDo.DTO
         public bool DaThanhLy { get => daThanhLy; set => daThanhLy = value; }
         
 
-        public SanPhamDTO(string MaSP, string TenLoai, string TenSP, float DinhGia,float GiaThanhLy, string MoTa, string MauSac , string HienTrang, string NhangHieu , bool QuaHan , bool DaChuoc , bool ThanhLy,bool DaThanhLy)
+        public SanPhamDTO(string MaSP, string TenLoai, string TenSP, double DinhGia, double GiaThanhLy, string MoTa, string MauSac , string HienTrang, string NhangHieu , bool QuaHan , bool DaChuoc , bool ThanhLy,bool DaThanhLy)
         {
             this.MaSP = MaSP;
             this.TenLoai = TenLoai;
@@ -59,8 +59,8 @@ namespace DoAnTTTT_QuanLyCuaHieuCamDo.DTO
             this.MaSP = row["MaSP"].ToString();
             this.TenLoai = row["TenLoai"].ToString();
             this.TenSP = row["TenSP"].ToString();
-            this.DinhGia = (float)Convert.ToDouble(row["DinhGia"]);
-            this.GiaThanhLy = (float)Convert.ToDouble(row["GiaThanhLy"]);
+            this.DinhGia = (double)Convert.ToDouble(row["DinhGia"]);
+            this.GiaThanhLy = (double)Convert.ToDouble(row["GiaThanhLy"]);
             this.MoTa = row["MoTa"].ToString();
             this.MauSac = row["MauSac"].ToString();
             this.HienTrang = row["HienTrang"].ToString();
